requirements.txt
asgiref==3.3.1
attrs==20.2.0
certifi==2020.6.20
chardet==3.0.4
click==7.1.2
configobj==5.0.6
Django==3.1.7
djangorestframework==3.12.2
drf-yasg==1.20.0
ecdsa==0.16.0
idna==2.10
importlib-metadata==3.7.2
iniconfig==1.1.1
packaging==20.4
pbkdf2==1.3
pip-tools==5.3.1
pluggy==0.13.1
psycopg2==2.8.6
py==1.9.0
pyaes==1.6.1
pyblake2==1.1.2
pycurl==7.43.0.6
pyparsing==2.4.7
pytest==6.1.1
python-dotenv==0.14.0
pytz==2021.1
requests==2.24.0
sentry-sdk==0.19.4
six==1.15.0
slick-bitcoinrpc==0.1.4
sqlparse==0.4.1
toml==0.10.1
typing-extensions==3.7.4.3
ujson==4.0.2
urllib3==1.25.10
zipp==3.4.1
